requirements.txt
slack_bolt
slack_sdk
flask
requests
python-dotenv